Edited for technical accuracy
I am an cryptocurrency and blockchain expert. According to my understanding of your system (which could be wrong) there are two main issues with this page.

You claim to have a "permissionless" system. However your ledger is stored on a central database with _trusted_ administrators. This is called a "permissioned" system.

*  I have corrected the text

You claim to have a blockchain. However your brikchain explorer does not show the hashes that would be required to identify and authenticate blocks. You seem to have a conventional centralised database.

For this reason I have have put the word 'blockchain' in quote marks wherever it appears.

OTHER
It is possible/likely that real blockchain technology would be a very good fit for the Ecobricks organisation. For example, brikcoins could have a real financial value and those who mint them could have votes on community proposals.

If you would like to talk about this you can contact me at james@trianglepm.org.